Adding an interactive to the generate_resume
