--- commands/skills.py
@@ -1,7 +1,0 @@
-import typer
-
-def add_skills():
-    """Add skills to the resume.yaml file."""
-    
-    typer.echo("Adding skills to resume.yaml...")
-    # Logic to add skills to resume.yaml would go here
