+++ resume.pdf
@@ -37,7 +37,7 @@ endobj
 endobj
 7 0 obj
 <<
-/Author (anonymous) /CreationDate (D:20250726134356-04'00') /Creator (ReportLab PDF Library - www.reportlab.com) /Keywords () /ModDate (D:20250726134356-04'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
+/Author (anonymous) /CreationDate (D:20250727092328-04'00') /Creator (ReportLab PDF Library - www.reportlab.com) /Keywords () /ModDate (D:20250727092328-04'00') /Producer (ReportLab PDF Library - www.reportlab.com) 
   /Subject (unspecified) /Title (untitled) /Trapped /False
 >>
 endobj
@@ -48,10 +48,10 @@ endobj
 endobj
 9 0 obj
 <<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 1237
+/Filter [ /ASCII85Decode /FlateDecode ] /Length 894
 >>
 stream
-Gau`Q9p=Qp&A@P9kZeb+"U9Il^t^gN1@2db"l;lF.#_uL,>4%B_buaAT9]N4bd%OS,rMP.PV%$mmappLKarRei]Ba5IL2]2G/<hoN55sn!5^1D_KYe14YH-qSDM%#\8^'g3prY"J1s-&l8[U/S*0G12V5FqNGq8D3AN?6k%=t75klB=9-RcFLkqFY,l@fb@krJiDC<YEns-Q@1-)JEp\+[_SJhLs($\Qa;$f)AMs'VD\XYcHI@.nBdqn_dnbEmu'(ABV5JG:O-R.\5'o!;1>/k/E'SSF#;D'c*J;?'"CmU)j(>1Po_>Y:-V1RTlH,/"HUQ@QGT9MVd7>YFsJIP71I,EX5buWW>9r63+iCbUbd7e:eUgd:F&kU2&VBZ:%<Ej2oQSDQ6Ou4G`YXL<AF2?)WJ[+9@mcYQiH]VHZkOtC9:8_9sS"M]\(D2Y2b*\VDJmi(UQms#Y1aA,3&5eROKVmBDk6gp0?>pYViFr#U1`9K^+",P%nnjb4_3Ba$a85Ek4G:%T'8F\A',LCY?X*/<g$d;9F)Pc2*kB%[-f)R^r*'mjqI;]m9L_gqdG)%j:gFa[9VJbLP>&"-r!:2\HmTHNp:jiS0sK8L?"h&Z;2t2?D9!0U8-oq6D[ee:VPp=rj-]k!OGj=RcI(>ql/3>R!>f3N7F4DM&Eb7C!=3<B\"Z:]BELZ?16YVC*`3m+W<&J#@Aa98.&nhnU8;G(+psa$?YEt,E:uV,M@.Op2AQTk8]cUH+lgM68Yfo_qA!5TW5&\i:/](lJjH&L9t0fn95=mg!\E?m/']>p8S%2p:7T:$)[qYe6XWDEb.@@t&p%@XQ/mcnV9Utu=f=sTaT`U1LuDE+s140YQ)#;$$GOLor-ls0>TnX<d:Y]VHlNU52_p%)lfGHl_fRV?JA7DHo?#0rKqh0*E:J?FeYgCYrq-STc#1_BU"d;CWBKsi/QEL'06RBBEnBX95s:P!"^T7;H@@QHo?hfUGUTqJQ\qt>A@"G;9]9rVRb8d:J]%)seHK1`I<,F<-ZCU6^3A0j6[qek*4)uoJ*&-e.pC#i6:IS]MW$Z--"'aF2,uWN8'TA^K!/hP]-qPL[t!#5WX\mO=g@ng)'uH7d=`:J`RELk8ITrfRi630U#2A@iGbruXjXpaL/rs`&P8h'078i[7^5`L3X?3JkIC<5S8*+.mB9sTgk>*$NH3?.B,hSncu_)[_".,IQJgk82]!/25tuu<E!h81#k.`s1q-Kf~>endstream
+Gau`QhbW5s&:XAWR)BfR5bW-rn`adW5u@s<(_^8BFsq^jmP_[]d)80JVFAm2bk2S^B]uYQREkQObdT43k]"^dPQIA.cocBcD*p8fpS[Tr_@cFX4aQS2/F@bDf5(+c3'?IeDU7W7R')u`::G=\4'pn.G)C@cB3NCbL#N*Lh[od;g5)Y`AkT3m\3Llc)&kP$]0_bO$jjD$I-*gHD%;dAC),sfYS]7NIMK\29"fs/ra06gBt_]mr;ffs4ELY4n`8i65Fi:&Y5Jr+Dm->Aao-#MD5_3/DuOTe3Q:)SAg#Q;NlH#aR=H$7!Vi%UjsVGW/g4eh-!Vc.W>L3-K3?Jg>eDd`jH^e1-74YX1'X%<F-T[?_<bFke2ZC,Qs!f4">3QT,Gl/P9<]6bU_509BYZ)O<[16@QahkkPESgGMUFi21G"K^9I?LA(4jVm)iC8!Jk)n9b,J2pYN%Wu'rohL]aan$MRK!P9iSJbA#Ui2dTCC.6EK;\^(U0>!G^4JEk3!YLAbM!4Ud59]Gur$e(l-a'Q]`(/4"=cj!VH8Z47[eaek3_RB8i\?\e/Z>QoAQj`DsP^[0h[:$4r0#^6DtKemS@c/Ue8#j2P>RL(NK^[19\n<S1_%VLtZ!.R,d1j\P41-5D<d$fI)Dk;)8A!<a3.r:sE).(PelhS7g29$U$*#fejSmnH/'\"'=rihNa'l&\!%[8@u+[J]";%3:gVsBLIH%EM,.@bH@aCc&?b:6H\Y8h^Tc5e@gS^NQ<+B7OuZ3Ds[8DKG=rAG<!4HQG0o"J3SPqBq&a^^8O%]U05nbYBm_AC(M,2/IAg0uQ?\LFtM@/kT3J,8HUn/I="rb;+dQsahP!"EpX`XF$*>2(0..=Mtf2(Kr6ha>T<A9L6JBftb\!FEUP:&~>endstream
 endobj
 xref
 0 10
@@ -68,7 +68,7 @@ xref
 trailer
 <<
 /ID 
-[<1f21d2888965965effc2ac808f99b808><1f21d2888965965effc2ac808f99b808>]
+[<9e29dac24ed45488c90b25da711c6230><9e29dac24ed45488c90b25da711c6230>]
 % ReportLab generated PDF document -- digest (http://www.reportlab.com)
 
 /Info 7 0 R
@@ -76,5 +76,5 @@ trailer
 /Size 10
 >>
 startxref
-2380
+2036
 %%EOF
