+++ resume.yaml
@@ -1,73 +1,79 @@
-login: octocat
-name: The Octocat
-company: '@github'
-blog: https://github.blog
+login: HollowCrusader
+name: AaronJoel
+company: Rochester Institute of Technology
+blog: ''
 email: null
 hireable: null
-bio: null
-location: San Francisco
-twitter_username: null
-followers: 18765
-following: 9
-created_at: '2011-01-25T18:44:36Z'
-updated_at: '2025-07-22T11:24:35Z'
-public_repos: 8
-public_gists: 8
+bio: DevOps wizard who wrangled code by day and shreds guitar by night.
+location: Limgrave, The Lands Between
+twitter_username: Aaroboo
+followers: 6
+following: 3
+created_at: '2017-04-28T13:29:22Z'
+updated_at: '2025-07-27T13:22:58Z'
+public_repos: 9
+public_gists: 0
 repos:
-  - name: boysenberry-repo-1
-    description: Testing
+  - name: ai-github-resume-builder
+    description: null
+    language: Python
+    stars: 0
+    forks_count: 0
+    topics: []
+    html_url: https://github.com/HollowCrusader/ai-github-resume-builder
+  - name: direct-file
+    description: Direct File
     language: null
-    stars: 346
-    forks_count: 23
+    stars: 0
+    forks_count: 0
     topics: []
-    html_url: https://github.com/octocat/boysenberry-repo-1
-  - name: linguist
-    description: Language Savant. If your repository's language is being reported
-      incorrectly, send us a pull request!
-    language: Ruby
-    stars: 571
-    forks_count: 216
-    topics: []
-    html_url: https://github.com/octocat/linguist
-  - name: test-repo1
+    html_url: https://github.com/HollowCrusader/direct-file
+  - name: grpc-microservices-in-go
     description: null
     language: null
-    stars: 335
-    forks_count: 28
+    stars: 0
+    forks_count: 0
     topics: []
-    html_url: https://github.com/octocat/test-repo1
-  - name: hello-worId
-    description: My first repository on GitHub.
+    html_url: https://github.com/HollowCrusader/grpc-microservices-in-go
+  - name: webrtc
+    description: Pure Go implementation of the WebRTC API
     language: null
-    stars: 574
-    forks_count: 228
+    stars: 0
+    forks_count: 0
     topics: []
-    html_url: https://github.com/octocat/hello-worId
-  - name: git-consortium
-    description: This repo is for demonstration purposes only.
-    language: null
-    stars: 444
-    forks_count: 124
-    topics: []
-    html_url: https://github.com/octocat/git-consortium
-  - name: octocat.github.io
+    html_url: https://github.com/HollowCrusader/webrtc
+  - name: DvnKiOS
     description: null
-    language: CSS
-    stars: 878
-    forks_count: 430
-    topics: []
-    html_url: https://github.com/octocat/octocat.github.io
-  - name: Spoon-Knife
-    description: This repo is for demonstration purposes only.
-    language: HTML
-    stars: 13126
-    forks_count: 152349
-    topics: []
-    html_url: https://github.com/octocat/Spoon-Knife
-  - name: Hello-World
-    description: My first repository on GitHub!
     language: null
-    stars: 3041
-    forks_count: 3276
+    stars: 0
+    forks_count: 0
     topics: []
-    html_url: https://github.com/octocat/Hello-World
+    html_url: https://github.com/HollowCrusader/DvnKiOS
+  - name: HollowCrusader
+    description: null
+    language: null
+    stars: 0
+    forks_count: 0
+    topics: []
+    html_url: https://github.com/HollowCrusader/HollowCrusader
+  - name: securemancala
+    description: null
+    language: PHP
+    stars: 0
+    forks_count: 0
+    topics: []
+    html_url: https://github.com/HollowCrusader/securemancala
+  - name: WineWorx
+    description: null
+    language: JavaScript
+    stars: 0
+    forks_count: 0
+    topics: []
+    html_url: https://github.com/HollowCrusader/WineWorx
+  - name: AllAccess
+    description: null
+    language: C#
+    stars: 0
+    forks_count: 0
+    topics: []
+    html_url: https://github.com/HollowCrusader/AllAccess
